Move to using trollop
Seems more flexible and better suited to the problem than Thor. Perhaps
slightly more code to write, but it seems easier to test as well as meaning
We can make less compromises (we can call our subcommands what we like, so
we can have slashes and other things in them).